sk6_studi_kasus_6/sk6/requirements.txt
flask
flask-cors
sqlalchemy
python-dotenv
requests
